Type EPOCH datetimes as Long in Kotlin and Scala
Both inferred a map value's type from the epoch value's width, so a
timestamp after 2038-01-19 was typed ``Int`` while its literal was a
``Long``, and nothing using the option compiled.  A map value's type
cannot depend on the value, so ``EPOCH`` is always a ``Long`` now and
its literal always carries the suffix that makes it one.

That also fixes a Kotlin annotation the old rule produced --
``val my_data: Long = 1705321800`` -- where the declared type and the
literal disagreed in the other direction.

Fixes #3965.

+++ tests/integration/cases/dict_all_scalar_types/Kotlin_heterogeneous_strategy_tuple_datetime_epoch@v1_9.kts
@@ -1,5 +1,5 @@
 import java.time.LocalDate
-data class Record0(val s: String, val i: Int, val f: Double, val b: Boolean, val n: Any?, val d: LocalDate, val dt: Int, val by: String)
+data class Record0(val s: String, val i: Int, val f: Double, val b: Boolean, val n: Any?, val d: LocalDate, val dt: Long, val by: String)
 val my_data = Record0(
     s = "string",
     i = 1,
@@ -7,6 +7,6 @@ val my_data = Record0(
     b = true,
     n = null,
     d = LocalDate.of(2024, 1, 15),
-    dt = 1705320000,
+    dt = 1705320000L,
     by = "48656c6c6f",
 )
