+++ tests/integration/cases/dict_all_scalar_types/Scala_heterogeneous_strategy_record_datetime_epoch@v3.scala
@@ -1,6 +1,6 @@
 import java.time.LocalDate
 object Fixture_dict_all_scalar_types_Scala_heterogeneous_strategy_record_datetime_epoch {
-case class Record0(s: String, i: Int, f: Double, b: Boolean, n: Any, d: LocalDate, dt: Int, by: String)
+case class Record0(s: String, i: Int, f: Double, b: Boolean, n: Any, d: LocalDate, dt: Long, by: String)
 val my_data = Record0(
     s = "string",
     i = 1,
@@ -8,7 +8,7 @@ val my_data = Record0(
     b = true,
     n = null,
     d = LocalDate.of(2024, 1, 15),
-    dt = 1705320000,
+    dt = 1705320000L,
     by = "48656c6c6f",
 )
 }
